src/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
google-cloud-aiplatform==1.71.0
google-adk==1.26.0
firebase-admin==6.5.0
langchain==0.3.7
langchain-community==0.3.7
langchain-text-splitters==0.3.2
pypdf==5.1.0
pydantic==2.9.2
python-dotenv==1.0.1
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
